% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowl18-harmonicity.R
\name{bowl18_harmonicity}
\alias{bowl18_harmonicity}
\title{Harmonicity}
\usage{
bowl18_harmonicity(x, tonic = 0L)
}
\arguments{
\item{x}{Chord to analyse,
expressed as an integerish vector of MIDI note numbers,
or more generally as any valid input to \code{\link[hrep:pi_chord]{pi_chord}}.}

\item{tonic}{(Integerish scalar, default = 0L)
Tonic to use when defining the just-tuned scale.}
}
\value{
(Numeric scalar)
The chord's harmonicity, defined as the proportion of harmonics
of the chord's fundamental frequency that coincide with the
harmonics of the chord tones.
}
\description{
Computes the harmonicity of a chord using the algorithm of
Bowling et al. (2018).
The algorithm assumes that chord pitches are precisely
tuned to the just-tuned scale provided by Bowling et al. (2018).
}
\examples{
bowl18_harmonicity(c(60, 64, 67))
bowl18_harmonicity("60 64 67")
bowl18_harmonicity(hrep::pi_chord(c(60, 64, 67)))
}
\references{
\insertRef{Bowling2018}{bowl18}
}
