% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowl18-harmonicity.R
\name{bowl18_harmonicity}
\alias{bowl18_harmonicity}
\alias{bowl18_harmonicity.default}
\alias{bowl18_harmonicity.pi_chord}
\title{Harmonicity}
\usage{
bowl18_harmonicity(x)

\method{bowl18_harmonicity}{default}(x)

\method{bowl18_harmonicity}{pi_chord}(x)
}
\arguments{
\item{x}{Chord to analyse.
The default method assumes that the chord is expressed
as a numeric vector of MIDI note numbers.
This can be made explicit by using the \code{\link[hrep:pi_chord]{pi_chord}}
function from the \code{hrep} package.}
}
\value{
(Numeric scalar)
The chord's harmonicity, defined as the proportion of harmonics
of the chord's fundamental frequency that coincide with the
harmonics of the chord tones.
}
\description{
Computes the harmonicity of a chord using the algorithm of
Bowling et al. (2018).
The algorithm assumes that chord pitches are precisely
tuned to the just-tuned scale provided by Bowling et al. (2018).
}
\examples{
bowl18_harmonicity(hrep::pi_chord(c(60, 64, 67)))
}
\references{
\insertRef{Bowling2018}{bowl18}
}
